move constants to common folder
