+++ packages/config/src/common/technology.ts
