--- packages/config/src/layer2s/common/assessCount.ts
@@ -1,10 +1,0 @@
-import { AssessCount } from '../types'
-
-export const subtractOne: AssessCount = (count) => count - 1
-
-export function subtractOneAfterBlockInclusive(
-  blockNumber: number,
-): AssessCount {
-  return (count: number, block: number) =>
-    block >= blockNumber ? count - 1 : count
-}
