--- packages/config/src/layer2s/common/constants.ts
@@ -1,1 +1,0 @@
-export const DANGER_DELAY_THRESHOLD_SECONDS = 60 * 60 * 12
