+++ packages/config/src/layer2s/common/index.ts
@@ -1,11 +1,2 @@
-export * from './assessCount'
-export * from './dataAvailability'
-export * from './exits'
-export * from './forceTransactions'
-export * from './milestones'
-export * from './newCryptography'
-export * from './operator'
-export * from './riskView'
+export * from './liveness'
 export * from './stages'
-export * from './stateCorrectness'
-export * from './technology'
